services/ollama-router/pyproject.toml
[project]
name = "ollama-router"
version = "0.1.0"
description = "Central LLM inference router for multiple Ollama instances"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "prometheus-client>=0.20",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
